Podlaczono MOPSO pod symulacj regulatora PID i obiektu 4 stopnia

+++ ssmodels.h
@@ -69,12 +69,12 @@ DECLARE_CONTINUOUS_DYNAMICAL_SYSTEM(SysClosedLoopUsWithPID)
 
 END_DECLARATION
 
-int CheckStabilityRouthHurwitz(std::vector<float> poly);
-int CheckStabilityOfMySystem(float Kp,float Ki,float Kd);
+int CheckStabilityRouthHurwitz(std::vector<double> poly);
+int CheckStabilityOfMySystem(double Kp,double Ki,double Kd);
 
-int GetStepResponse(ContinuousDynamicalSystem*sys,float Kp,float Ki,float Kd,float time,float dt,QVector<QPair<float,float>>& data);
-int GetImpulseResponse(ContinuousDynamicalSystem*sys,float Kp,float Ki,float Kd,float time,float dt,QVector<QPair<float,float>>& data);
+int GetStepResponse(ContinuousDynamicalSystem*sys,double Kp,double Ki,double Kd,double time,double dt,QVector<QPair<double,double>>& data);
+int GetImpulseResponse(ContinuousDynamicalSystem*sys,double Kp,double Ki,double Kd,double time,double dt,QVector<QPair<double,double>>& data);
 
-void GetStepResponseCharacteristics(QVector<QPair<float,float>>& data,float &MaxOvershoot,float &RiseTime,float &SettlingTime);
+void GetStepResponseCharacteristics(QVector<QPair<double,double>>& data,double &MaxOvershoot,double &RiseTime,double &SettlingTime);
 
 #endif // EXAMPLE_H
